Remove resident rows when deselected
When deselecting a house on the map, hide the resident rows from being
displayed in the table. This prevents the resident data from being shown
even when the house is no longer selected. Allow select, residents, and
house props and variables to be null to signify when nothing is selected
based on what values are give or set.